Restructure files, fix binary search and add docstrings
* Add algorithms folder and tests folder
* Fix stop criteria and binary search recursive calls
* Add docstring for quick_sort_m3 and improve quick_sort_docstring
