+++ algorithms/tower_of_hanoi.py
